feat: dashboard studios redesign with realtime inbox, search, and media tools
Redesign the authenticated dashboard around the consolidated Pulse Mosaic
direction: shared DashboardWorkspace layout, colored sidebar icons, red theme
tokens, motion-safe animations, and localized member and admin pages.

Add production features behind permissions:
- Ctrl+K workspace search backed by DashboardSearchController (owned
  conversations, media, tools, templates; literal wildcard escaping;
  uuid-safe matching on PostgreSQL).
- Realtime inbox over Reverb private channels (NotificationChanged event,
  channel authorization, NotificationContext with auth-failure handling).
- Separate image, video, and audio studios: video reference images and Pro
  mode (2x tokens), Kokoro voiceover and Stable Audio music via fal.
- yt-dlp downloads and FFmpeg conversions through sandboxed workers
  (MediaToolService, ProcessMediaToolJob, Range-served private assets).

Include migrations, seeders, locale keys, PostgreSQL test guard, e2e and
feature contract tests, and local runtime scripts.

+++ tests/Feature/ReferralProgramTest.php
@@ -211,7 +211,7 @@ class ReferralProgramTest extends TestCase
         $this->assertSame(now()->addDays(12)->toISOString(), $referred->fresh()->expires_at->toISOString());
         $this->assertDatabaseCount('referral_rewards', 2);
         $this->assertSame(2, ReferralReward::distinct()->count('reference'));
-        $this->assertSame(10, ReferralReward::sum('days'));
+        $this->assertSame(10, (int) ReferralReward::sum('days'));
         $this->assertDatabaseCount('notifications', 2);
         $this->assertEqualsCanonicalizing(
             [$referrer->id, $referred->id],
@@ -256,6 +256,7 @@ class ReferralProgramTest extends TestCase
         $this->assertTrue($referredExpiryAfterReward->copy()->addDay()->equalTo($referred->fresh()->expires_at));
         $this->assertDatabaseCount('referral_rewards', 2);
     }
+
     public function test_manual_duration_grant_does_not_consume_first_paid_referral_reward(): void
     {
         $admin = User::factory()->create(['role' => 'admin']);
